Moved test into the MinerTests class.
