+++ src/Stratis.Bitcoin.IntegrationTests/MinerTests.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using FluentAssertions;
 using Microsoft.Extensions.Logging;
 using NBitcoin;
 using NBitcoin.DataEncoders;
@@ -18,6 +19,7 @@ using Stratis.Bitcoin.Features.Consensus.Rules;
 using Stratis.Bitcoin.Features.MemoryPool;
 using Stratis.Bitcoin.Features.MemoryPool.Fee;
 using Stratis.Bitcoin.Features.Miner;
+using Stratis.Bitcoin.IntegrationTests.EnvironmentMockUpHelpers;
 using Stratis.Bitcoin.P2P;
 using Stratis.Bitcoin.P2P.Peer;
 using Stratis.Bitcoin.P2P.Protocol.Payloads;
@@ -608,6 +610,31 @@ namespace Stratis.Bitcoin.IntegrationTests
             context.chain.SetTip(new BlockHeader { HashPrevBlock = context.chain.Tip.HashBlock, Time = Utils.DateTimeToUnixTime(context.chain.Tip.GetMedianTimePast()) + 512 });
             context.chain.SetTip(new BlockHeader { HashPrevBlock = context.chain.Tip.HashBlock, Time = Utils.DateTimeToUnixTime(context.chain.Tip.GetMedianTimePast()) + 512 });
             Assert.True(tx.IsFinal(context.chain.Tip.GetMedianTimePast().AddMinutes(2), context.chain.Tip.Height + 2)); // Locktime passes 2 min later
+        }
+
+        [Fact]
+        public void GetProofOfWorkRewardForMinedBlocksTest()
+        {
+            using (NodeBuilder builder = NodeBuilder.Create())
+            {
+                var node = builder.CreateStratisPowNode();
+                builder.StartAll();
+                node.NotInIBD();
+
+                node.SetDummyMinerSecret(new BitcoinSecret(new Key(), node.FullNode.Network));
+
+                node.GenerateStratisWithMiner(10);
+                node.GetProofOfWorkRewardForMinedBlocks(10).Should().Be(Money.Coins(500));
+
+                node.GenerateStratisWithMiner(90);
+                node.GetProofOfWorkRewardForMinedBlocks(100).Should().Be(Money.Coins(5000));
+
+                node.GenerateStratisWithMiner(100);
+                node.GetProofOfWorkRewardForMinedBlocks(200).Should().Be(Money.Coins(8725));
+
+                node.GenerateStratisWithMiner(200);
+                node.GetProofOfWorkRewardForMinedBlocks(400).Should().Be(Money.Coins((decimal)12462.50));
+            }
         }
 
         //NOTE: These tests rely on CreateNewBlock doing its own self-validation!
