+++ src/Stratis.Bitcoin.IntegrationTests/UtilTests.cs
@@ -1,9 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using FluentAssertions;
-using NBitcoin;
-using Stratis.Bitcoin.IntegrationTests.EnvironmentMockUpHelpers;
 using Stratis.Bitcoin.Utilities;
 using Xunit;
 
@@ -44,31 +41,6 @@ namespace Stratis.Bitcoin.IntegrationTests
             task.Wait();
 
             Assert.True(IsSequential(collector.ToArray()));
-        }
-
-        [Fact]
-        public void GetProofOfWorkRewardForMinedBlocksTest()
-        {
-            using (NodeBuilder builder = NodeBuilder.Create())
-            {
-                var node = builder.CreateStratisPowNode();
-                builder.StartAll();
-                node.NotInIBD();
-
-                node.SetDummyMinerSecret(new BitcoinSecret(new Key(), node.FullNode.Network));
-
-                node.GenerateStratisWithMiner(10);
-                node.GetProofOfWorkRewardForMinedBlocks(10).Should().Be(Money.Coins(500));
-
-                node.GenerateStratisWithMiner(90);
-                node.GetProofOfWorkRewardForMinedBlocks(100).Should().Be(Money.Coins(5000));
-
-                node.GenerateStratisWithMiner(100);
-                node.GetProofOfWorkRewardForMinedBlocks(200).Should().Be(Money.Coins(8725));
-
-                node.GenerateStratisWithMiner(200);
-                node.GetProofOfWorkRewardForMinedBlocks(400).Should().Be(Money.Coins((decimal)12462.50));
-            }
         }
     }
 }
